4689 using regex for vars under cursor
